+++ sshconnect2.c
@@ -340,7 +340,7 @@ struct cauthctxt {
 #ifdef GSSAPI
 	/* gssapi */
 	gss_OID_set gss_supported_mechs;
-	u_int mech_tried;
+	size_t mech_tried;
 #endif
 	/* pubkey */
 	struct idlist keys;
